backend-ia/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
opencv-python
ultralytics
torch
supabase
python-dotenv
python-jose[cryptography]